Move inertia tensor functions into their own file
